refactor: mypage type 관리
